Collaborative Notebook Rework Version 3: - Reformatted and reworked the code for better readability & modularity. - All CF algorithms are now in a single notebook. - Added a saving process for hyperparameters after tuning for easy re-training. - Added sections for formatting and saving training results.
(Keep in mind that LightFM models have contiounsly caused kernel crashes
on windows systems and have only been able to run on linux for me.
This is why the data loading and training process for LightFM is
conditionally seperated from the rest of the models.)

May or may not add SVDpp to the notebook later on. But for now it's
pretty much done.